// src/shared/models/domain/client/job.postings.domain.ts
export interface IClientJobPostingCompany {
  id: string;
  name: string;
  description: string;
  companyType: string; // e.g. "ENTERPRISE"
  industry: string; // e.g. "EDUCATION"
  size: string; // e.g. "OVER_THOUSAND"
  stage: string; // e.g. "ENTERPRISE"
  foundedYear: number;
  website: string;
  benefits: string[];
  contactEmail: string;
  contactPhone: string;
  contactName: string;
  address: string;
  city: string;
  state: string;
  zipCode: string;
  country: string;
  remarks: string | null;
  status: string; // e.g. "UNVERIFIED"
  logo: string | null;
}
// ...
export interface IClientJobPosting {
  id: string;
  clientId: string;
  applications: IClientJobApplication[];
  candidateShortlists?: IClientCandidateShortlist[];
  title: string;
  description: string;
  jobType: WorkTypeEnum;
  jobCommitment: WorkCommitmentEnum;
  jobSchedule: WorkScheduleEnum;
  industry: CompanyIndustryEnum;
  totalExperience: number;
  department?: string;
  teamSize?: number;
  reportingTo?: string;
  hiring_manager_email?: string;
  status: JobPostingStatusEnum;
  applicationDeadline?: Date;
  availableFrom?: Date;
  numberOfOpenings: number;
  applicationUrl?: string;
  isFeatured: boolean;
  numberOfViews: number;
  numberOfApplications: number;
  minSalary: number;
  maxSalary: number;
  salaryCurrency: string;
  equity: boolean;
  responsibilities: string[];
  benefits: string[];
  tags: string[];
  isRemote: boolean;
  preferredUniversities: string[];
  preferredDegrees: string[];
  preferredLocations: string[];
  preferredIndustries: string[];
  requiredSkills: string[];
  preferredSkills: string[];
  isPublished: boolean;
  createdAt: Date;
  updatedAt: Date;
  company?: IClientJobPostingCompany;
}